move everything into socket service
